requirements.txt
asgiref==3.5.2
bleach==5.0.1
coverage==6.4.1
Django==4.0.5
django-appconf==1.0.5
django-cors-headers==3.13.0
django-imagekit==4.1.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
flake8==4.0.1
mccabe==0.6.1
pilkit==2.0
Pillow==9.2.0
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
webencodings==0.5.1
